Move TS code from putting_together.py to custom.ts
